third_party/blink/web_tests/platform/generic/fast/events/scale-and-scroll-iframe-body-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS frame.contentDocument.scrollingElement.scrollTop is 100
PASS frame.contentDocument.scrollingElement.scrollLeft is 100
PASS frame.contentWindow.scrollX is 100
PASS frame.contentWindow.scrollY is 100
